requirements.txt
torch==1.6.0
torchvision==0.7.0
tensorboardx>=2.0
scikit-learn>=0.22.1
gitpython
ordered-set
termcolor
wget